Add files via upload
Navigation done! With backgrounds photos and footer! Also in checkout page its already a lot of the labels done. I' m sending screenshots to get an idea and I will try to push the folder on github. CSS and HTML pages are created and work each other. 
(The folder is also uploaded in my fork section)